+++ packages/polaris-viz/src/utilities/getTrendIndicatorData.ts
@@ -1,11 +1,14 @@
-import type {MetaDataTrendIndicator} from '../../../SimpleBarChart';
-import {estimateTrendIndicatorWidth} from '../../../TrendIndicator';
+import type {MetaDataTrendIndicator} from 'types';
+
+import {estimateTrendIndicatorWidth} from '../components';
 
 export function getTrendIndicatorData(
   trendMetadata: MetaDataTrendIndicator | undefined,
 ) {
   if (trendMetadata != null) {
-    const {totalWidth} = estimateTrendIndicatorWidth(`${trendMetadata.value}`);
+    const {totalWidth} = estimateTrendIndicatorWidth(
+      `${trendMetadata.value || ''}`,
+    );
 
     return {
       trendIndicatorProps: trendMetadata,